rename dashboarditem.options to .config
