+++ src/router/routes.ts
@@ -5,7 +5,7 @@ export default [
     children: [
       { path: '', component: () => import('@/pages/IndexPage.vue') },
       { path: '/settings', component: () => import('@/pages/SettingsPage.vue') },
-      { path: '/blocks', component: () => import('@/pages/BlocksDashboardPage.vue') },
+      { path: '/blocks', component: () => import('@/pages/BlocksPage.vue') },
 
       // dynamic pages
       { path: '/dashboard/:id', component: () => import('@/pages/DashboardPage.vue') },
